Remove trailing blank line from index.js
